edit

+++ src/main/java/pro/sky/todolist/service/LabelService.java
@@ -3,12 +3,16 @@ package pro.sky.todolist.service;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 import pro.sky.todolist.dto.LabelDto;
-import pro.sky.todolist.exception.LabelIsPresentException;
-import pro.sky.todolist.exception.LabelNotFoundException;
+import pro.sky.todolist.dto.UserDto;
+import pro.sky.todolist.exception.*;
 import pro.sky.todolist.mapper.LabelMapper;
 import pro.sky.todolist.model.Label;
+import pro.sky.todolist.model.Task;
+import pro.sky.todolist.model.User;
 import pro.sky.todolist.repository.LabelRepository;
+import pro.sky.todolist.repository.TaskRepository;
 
+import java.util.List;
 import java.util.Optional;
 
 import static org.hibernate.validator.internal.util.Contracts.assertTrue;
@@ -16,10 +20,13 @@ import static org.hibernate.validator.internal.util.Contracts.assertTrue;
 @Service
 public class LabelService {
     private final LabelRepository labelRepository;
+    private final TaskRepository taskRepository;
+
     private  final LabelMapper labelMapper;
 
-    public LabelService(LabelRepository labelRepository, LabelMapper labelMapper) {
+    public LabelService(LabelRepository labelRepository, TaskRepository taskRepository, LabelMapper labelMapper) {
         this.labelRepository = labelRepository;
+        this.taskRepository = taskRepository;
         this.labelMapper = labelMapper;
     }
     @Transactional
@@ -45,11 +52,16 @@ public class LabelService {
         labelMapper.enrichLabel(labelDto, label);
         return labelMapper.toDto(labelRepository.save(label));
     }
-   // @Transactional
+    @Transactional
     public LabelDto delete(long id) {
         Label label = labelRepository.findById(id)
                 .orElseThrow(() -> new LabelNotFoundException(id));
-        labelRepository.delete(label);
+        List<Task> tasks = taskRepository.findTasksByLabelId(id);
+        if (!tasks.isEmpty()) {
+           throw new LabelHasTasksException(id);
+        } else
+            labelRepository.delete(label);
         return labelMapper.toDto(label);
     }
+
 }
